requirements_ref.txt
aenum==3.1.16
aiodns==3.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.0
aniso8601==8.0.0
annotated-types==0.7.0
ansicon==1.89.0
anthropic==0.75.0
anyio==4.11.0
async-generator==1.10
asyncio==4.0.0
attrdict==2.0.1
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
backtrader==1.9.78.123
banks==2.2.0
base58==2.1.1
bcrypt==5.0.0
beartype==0.22.6
beautifulsoup4==4.14.3
bitarray==3.8.0
blessed==1.25.0
boto3==1.41.5
botocore==1.41.5
build==1.3.0
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.3.7
chuk-ai-session-manager==0.7.1
chuk-llm==0.14.3
chuk-mcp==0.9
chuk-mcp-client-oauth==0.3.5
chuk-sessions==0.5.2
chuk-term==0.1.3
chuk-tool-processor==0.11.2
ckzg==2.1.5
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contexttimer==0.3.3
contourpy==1.3.3
cryptography==46.0.3
csrf==0.1b1
curl_cffi==0.13.0
cycler==0.12.1
cyclopts==4.3.0
cytoolz==0.12.3
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.3
dotenv==0.9.9
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
eth-abi==2.2.0
eth-account==0.4.0
eth-hash==0.7.1
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
eval_type_backport==0.3.0
exceptiongroup==1.3.1
faiss-cpu==1.13.1
fast-json==0.3.2
fastmcp==2.13.1
filelock==3.20.0
filetype==1.2.0
finnhub-python==2.4.26
firecrawl-py==4.11.0
flake8==7.3.0
Flask==1.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
flatbuffers==25.9.23
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-genai==1.52.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
griffe==1.15.0
groq==0.36.0
grpcio==1.76.0
h11==0.16.0
hdwallet==2.1.1
hexbytes==0.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
ibm-cos-sdk==2.14.3
ibm-cos-sdk-core==2.14.3
ibm-cos-sdk-s3transfer==2.14.3
ibm_watsonx_ai==1.4.7
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.13.0
invoke==2.2.1
ipfshttpclient==0.7.0
itsdangerous==1.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
kazoo==2.8.0
keyring==25.7.0
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.1.3
langchain-anthropic==1.3.0
langchain-core==1.2.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
lark==1.3.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index-agent-openai==0.4.12
llama-index-cli==0.5.3
llama-index-core==0.12.52.post1
llama-index-embeddings-huggingface==0.5.5
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-ollama==0.6.2
llama-index-llms-openai==0.4.7
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==1.3.0
llama-parse==0.6.54
loguru==0.7.3
lomond==0.3.3
lru-dict==1.4.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
marshmallow-sqlalchemy==0.28.0
matplotlib==3.10.8
mccabe==0.7.0
mcp==1.22.0
mcp-cli==0.11.1
mdurl==0.1.2
mistralai==1.9.11
mmh3==5.2.0
mnemonic==0.21
more-itertools==10.8.0
morphys==1.0
mpmath==1.3.0
multiaddr==0.1.0
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
nest-asyncio==1.6.0
netaddr==1.3.0
networkx==3.6.1
nltk==3.9.2
nodeenv==1.9.1
numpy==2.3.5
nv==0.0.7
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==1.109.1
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.4
ormsgpack==1.12.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
parsimonious==0.8.1
pathable==0.4.4
pathvalidate==3.3.1
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.0
posthog==5.4.0
pre_commit==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
py==1.11.0
py-cid==0.3.1
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==2.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycares==4.11.0
pycodestyle==2.14.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymultihash==0.8.2
PyMySQL==1.0.2
pyparsing==3.2.5
pypdf==5.9.0
pyperclip==1.11.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyrsistent==0.20.0
pythclient==0.1.4
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==14.2.0
rich-rst==1.3.2
rlp==1.2.0
rpds-py==0.29.0
rsa==4.9.1
s3transfer==0.15.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
sh==2.2.2
shellingham==1.5.4
simplejson==3.17.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.45
SQLAlchemy-Utils==0.42.1
srsly==2.5.2
sse-starlette==3.0.3
starlette==0.50.0
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thefuzz==0.22.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toolz==1.1.0
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
trio==0.32.0
trio-typing==0.10.0
tronapi==3.1.6
trx-utils==1.0.4
ty==0.0.1a34
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.3.0
uuid==1.30
uuid_utils==0.12.0
uv==0.9.13
uvicorn==0.38.0
varint==1.0.2
virtualenv==20.35.4
watchfiles==1.1.1
wcwidth==0.2.14
web3==5.8.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==1.0.1
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
yfinance==0.2.66
zipp==3.23.0
zstandard==0.25.0
